requirements.txt
flask
ultralytics
requests
Pillow
twilio
google-cloud-storage
uuid
numpy
opencv-python
Werkzeug
